refactor: replace 'isFailedVar' checks with exceptions
Old style 'isFailedVar' was used for CppDbg with GDB debugger. But after
the release of the first version, a lot of time has passed: logic become
more complex and extension must support different debuggers.

So using exceptions instead of C-style error check is necessary to be
able to scale the extension and write more clean code.