feat: cache typedefs.list in .vscode folder
After processing typedefs.list was saved in /tmp folder. This becomes an
issue i.e. when working with multiple versions of pg, when /tmp
mounted in tmpfs or when /tmp is cleaned up. So better idea is to store
this file locally - in .vscode folder which is perfect decision for
caching.