feat: 'typedefs' config param specifies additional typedefs.list
'typedefs' member in configuration file specified path to a ready to
use typedefs.list file. This is not very convinient because developers
can have miriads contribs in development and they have to update it
constantly.

Now, this member can be either string or list of strings each of them is
a path to custom typedefs.list which is just added to builtin
typedefs.list.