chore: find pg_config to build pg_bsd_indent in it's directory
To build pg_bsd_indent pg_config is required. Previously there was
prompt to enter path to it, but it's not very convinient and moreover
it's path must be already known - src/bin/pg_config/pg_config, so there
is no need to force user prompt fs path.

Also, prompt with asking to build pg_bsd_indent is removed - it's clear
that you want to build it if formatting is requested.