feat: show bitmask members as enum values instead of numbers
Some integer fields in structures acts not like numberic fields, but as
enum values (bitmask OR) and/or store integer values (apply bitmask to
get value).
Often this used in storage structures (i.e. t_infomask of heap tuple),
so it will be useful to show all such information instead of plain
numbers (for humans it is not informative).

This commit adds such feature - show enum values with field values, but
only for builtin members (no config support yet).

Also, there is limitation imposed by debug symbols - MACRO definitions
added to them only if '-g3' is enabled (which is not with default
'--enable-debug'), so to handle such situations we add macro numeric
value (as it is defined in source code). But note that major updates can
break everything and shown values will lead you down the wrong path.
Builtin values added from 17.4 version.