feat!: remove config file versioning
Extension is constantly evolving, so it's versioning was the method to
avoid breaking during updates. But it brings more pain, because not as
much people actually cares about version and (secondly) there are not so
many people that use this config file.

So removing versioning will make the code cleaner and more easy to think
about.