fix: array members not shown as array with codelldb debugger
Check for pointer type behaved incorrectly when pointer to base type was
passed, i.e. 'int *'. This was because when there was a pointer is type
then I checked it was a struct, but no checks that it was base type.
Now, this check is too dumb, so must pass.