feat: use pgindent instead of pg_bsd_indent for formatting
pg_bsd_indent was used because it seemed to be more performant because
pgindent performed some logic with temporary files. But this approach
requires to keep my code up to date with upstream and more over
different pg versions have different formatting rules.

So I am switching to using pgindent instead of copying it's behaviour.