refactor: add 'declaredType' member to 'Variable'
Single 'type' member is not enough for complex hierarchy, because we
have to constantly compute effective types and keep in mind which type
to show to user, so we just add separate member 'declaredType' - it
stores type that was originaly declared. Now, 'type' member will store
effective type and we will work only with it.