refactor: use single global logger
ILogger interface had to be passed to each object that want to log
anything. This slighly complicates architecture and forces to pass
logger everywhere.

But this is a VS Code extension with integration tests only (no unit
tests), so there is no need to mock logger.

Thus logger singleton is introduced. It lives in separate file and each
file has to just type

`import { Log as logger } from './logger'`

which in code looks like there no singleton at all.