feat: create .gitignore when bootstrapping extension
This can be handy when you decide to init repo and then move files to gh
repo, but for now extension runs only if contrib directory did not
exist, but when you first create gh repo, then it might be already
initialized and extension will not do anything.

If there will be such issue to add bootstraping to initialized repo,
then I will add it, but so far there is no such task.