feat!: remove custom nodetag files setting
This setting was added to keep track of new NodeTags, but file parsing
adds impact on performance and rarely used. It is a better idea to move
this to configuration file.

For now this setting is not removed from VS Code configuration, so not
error messages will be shown. Maybe I will find another way to use this.