perf: inline getStructNameRange into getStructNameFromType
getStructNameRange searches bounds of type name, but to do this it
returns array of 2 elements: start + end. But as it invoked oftenly this
can create lots of intermediate arrays and waste memory.

This function was inlined into getStructNameFromType and now
substituteStructName just invokes getStructNameFromType and then
'replace' with given typename. This is more optimal, because
substituteStructName invoked more rare that first.