test: run all debugger tests in a single variables.test.ts invocation
Variables tests require to know which debugger to use, but running tests
with single debugger is slow and uncomfortable.

Now, variables.test.ts will run each tests for each of debugger in
single run. That is run.sh invoked only once, thus increasing test
speed.

There is still an unpleasant situation - Extension Test Runner vscode
extension shows variables tests as skipped while they are not and I do
not know how to fix it. Anyway tests are run, but it will be more
convenient to see each test in a separate tree item with it's status.