feat: generalize config file records for HTAB and List ptr custom
Now they are in same format as internal represenatation: type, parent,
member - no specialization for member or variable.

Compatibility with previous format is saved, but schema.json changed to
new version without syntax help.

This is required to more convenient updating of configuration file in
callbacks (for now this is single update after new NodeTags found).