docs: initial files for static site on GitHub Pages
Extension has rich functionality, so it should have a good
documentation so users can understand how to use it.

For this documentation will be updated and stored in 'docs' directory.
Also, there will be a static web-site in 'Read-the-docs' format hosted
on GitHub Pages.

This commit adds first pages with configuration file explaining.