chore: extract all type properties instead of distinct function check
In Variable.create to check properties of type were used many separate
function, but this is not very convenient, because they performed many
redundant checks and (what is even worse) if we wanted to perform some
small changes in method of 1 debugger class then the whole sequence may
be disrupted and we have to change it, but it breaks checks of another.

This patch adds single entrypoint for extracting information about
variable - IDebuggerFacade.extractVariableProperties, which just
returns special object with all extracted knowledge of the given var.